backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
python-dotenv
httpx
cachetools
anthropic
pytest
pytest-asyncio
pytest-mock
